+++ dal/src/main/java/com/timevale/forward/dal/condition/PersonListCondition.java
@@ -29,5 +29,5 @@ public class PersonListCondition {
     /**
      * 人员类型:0项目-产品经理，1项目-项目成员，20产品需求-抄送人，30业务需求-抄送人
      */
-    private Byte type;
+    private Integer type;
 }
