+++ dal/src/main/java/com/timevale/forward/dal/condition/ProjectListCondition.java
@@ -29,7 +29,7 @@ public class ProjectListCondition extends QueryBase {
     /**
      * 优先级:0(P0),1(P1),2(P2)
      */
-    private List<Byte> priorities;
+    private List<Integer> priorities;
 
     /**
      * 业务域
@@ -44,12 +44,12 @@ public class ProjectListCondition extends QueryBase {
     /**
      * 项目类型:0产品研发项目,1技术优化项目,2日常迭代
      */
-    private List<Byte> types;
+    private List<Integer> types;
 
     /**
      * 项目状态:0待启动,10规划中,20研发中,30测试中,40已发布,50已暂停,60已作废
      */
-    private List<Byte> status;
+    private List<Integer> status;
 
     /**
      * 项目经理
