+++ dal/src/main/java/com/timevale/forward/dal/entity/ProductDemandDO.java
@@ -17,7 +17,7 @@ public class ProductDemandDO extends BaseDO {
     /**
      * 优先级:0(P0),1(P1),2(P2),3(P3)
      */
-    private Byte priority;
+    private Integer priority;
     /**
      * 产品线
      */
@@ -25,12 +25,12 @@ public class ProductDemandDO extends BaseDO {
     /**
      * 类型:0新增功能,1功能迭代,2体验优化,3技术需求,4安全需求
      */
-    private Byte type;
+    private Integer type;
 
     /**
      * 0待排期,10已列入项目,20项目进行中,30已完成上线,40已暂停,50已作废
      */
-    private Byte status;
+    private Integer status;
     /**
      * 负责人
      */
