+++ facade/src/main/java/com/timevale/forward/facade/api/request/ProjectAddReq.java
@@ -21,13 +21,13 @@ public class ProjectAddReq extends BaseReq {
     private String name;
 
     @ApiModelProperty("优先级:0(P0),1(P1),2(P2)")
-    private Byte priority;
+    private Integer priority;
 
     @ApiModelProperty("产品线")
     private List<Long> productLineIds;
 
     @ApiModelProperty("项目类型:0产品研发项目,1技术优化项目,2日常迭代")
-    private Byte type;
+    private Integer type;
 
     @ApiModelProperty("项目经理")
     private PersonAddReq pm;
